requirements.txt
fastapi==0.115.7
loguru==0.7.3
tortoise-orm==0.24.0
uvicorn==0.34.0
gunicorn==23.0.0
pydantic==2.10.6
aiogram==3.17.0
jinja2==3.1.5
asyncpg==0.30.0
python-dotenv==1.0.1
pytest==8.3.4
python-jose==3.3.0
python-multipart==0.0.20
aerich==0.2.4
httpx==0.27.2
bcrypt==4.3.0